fix: added way to set "cedana path" in code
when running tests, if you had a config file saved in "~/.cedana", you would have tests fail.  the tests that failed were those that checked for a specific error if that file didn't exist.  this commit allows a user, in code (maybe this could be a future cli feature?) to set the path of cedana config file only.  this allows us to create a non-conflicting cedana path to use with tests (or other uses).